requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
Django==5.1.1
django-celery-results==2.5.1
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
durationpy==0.9
eventlet==0.39.0
fastapi==0.115.12
frozenlist==1.5.0
google-auth==2.36.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
inflection==0.5.1
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
kubernetes==31.0.0
langchain==0.3.17
langchain-anthropic==0.3.5
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-redis==0.1.2
langchain-text-splitters==0.3.5
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-checkpoint-sqlite==2.0.3
langgraph-sdk==0.1.51
langsmith==0.3.5
marshmallow==3.26.1
ml-dtypes==0.4.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.59.7
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==12.1.1
prompt_toolkit==3.0.50
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.9.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==2.7.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
redisvl==0.3.9
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.22.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.1
tabulate==0.9.0
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.18.3
zstandard==0.23.0
uWSGI==2.0.26
django-storages[boto3]==1.14.4
